# [Performance] Approximate the ideal number of reports to render on first render