--- a/src/components/InvertedFlatList/index.android.js
+++ b/src/components/InvertedFlatList/index.android.js
@@ -1,18 +1,11 @@
 import React, {forwardRef} from 'react';
-import PropTypes from 'prop-types';
 import BaseInvertedFlatList from './BaseInvertedFlatList';
 
-const propTypes = {
-    /** The initial number of items to render */
-    initialNumToRender: PropTypes.number.isRequired,
-};
-
-const InvertedFlatList = forwardRef((props, ref) => (
+export default forwardRef((props, ref) => (
     <BaseInvertedFlatList
         // eslint-disable-next-line react/jsx-props-no-spreading
         {...props}
         ref={ref}
-        initialNumToRender={props.initialNumToRender}
 
         // Remove clipped subviews helps prevent avatars and attachments from eating up excess memory on Android. When
         // we run out of memory images stop appearing without any warning.
@@ -20,6 +13,3 @@ const InvertedFlatList = forwardRef((props, ref) => (
         shouldRemoveClippedSubviews
     />
 ));
-
-InvertedFlatList.propTypes = propTypes;
-export default InvertedFlatList;