--- a/src/components/InvertedFlatList/index.ios.js
+++ b/src/components/InvertedFlatList/index.ios.js
@@ -6,10 +6,5 @@ export default forwardRef((props, ref) => (
         // eslint-disable-next-line react/jsx-props-no-spreading
         {...props}
         ref={ref}
-
-        // Remove clipped subviews helps prevent avatars and attachments from eating up excess memory on Android. When
-        // we run out of memory images stop appearing without any warning.
-        // eslint-disable-next-line react/jsx-props-no-multi-spaces
-        shouldRemoveClippedSubviews
     />
 ));