<!-- If necessary, assign reviewers that know the area or changes well. Feel free to tag any additional reviewers you see fit. -->

### Details
<!-- Explanation of the change or anything fishy that is going on -->
Calculated the ideal number of reports to render in the first render based on the screen height and the minimum height of a report

### Fixed Issues
<!-- Please replace GH_LINK with the link to the GitHub issue this Pull Request is fixing -->
$ https://github.com/Expensify/App/issues/4258

### Tests
<!--- 
Add a numbered list of manual tests you performed that validates your changes work on all platforms, and that there are no regressions present.
Add any additional test steps if test steps are unique to a particular platform.
Manual test steps should be written so that your reviewer can repeat and verify one or more expected outcomes in the development environment.

For example:
1. Click on the text input to bring it into focus
2. Upload an image via copy paste
3. Verify a modal appears displaying a preview of that image 
--->
| Platform | initialNumToRender  | Average time to switch reports |
| ------------- | ------------- | ------------- |
| Android (Release) |  Fixed (20)  | 1,899 ms  |
| Android (Release) |  Ideal (10)  | 1,236 ms  |
| Web - Chrome (Dev) |  Fixed (20)  | 3,795.2 ms  |
| Web - Chrome (Dev) |  Ideal (7)  | 1,336.2 ms  |
| iOS - iPhone SE (Dev) |  Fixed (20)  | 7016 ms  |
| iOS - iPhone SE (Dev) |  Ideal (10)  | 3,888.6 ms  |

### QA Steps
<!--- 
Add a numbered list of manual tests that can be performed by our QA engineers on the staging environment to validate that your changes work on all platforms, and that there are no regressions present.
Add any additional QA steps if test steps are unique to a particular platform.
Manual test steps should be written so that the QA engineer can repeat and verify one or more expected outcomes in the staging environment.

For example:
1. Click on the text input to bring it into focus
2. Upload an image via copy paste
3. Verify a modal appears displaying a preview of that image 
--->

### Tested On

- [x] Web
- [x] Mobile Web
- [x] Desktop
- [x] iOS
- [x] Android

### Screenshots
<!-- Add screenshots for all platforms tested. Pull requests won't be merged unless the screenshots show the app was tested on all platforms.-->

#### Web
<!-- Insert screenshots of your changes on the web platform-->

https://user-images.githubusercontent.com/53711423/127908696-17e459d8-0d7c-4553-bec1-3827aa8bb5c1.mov


#### Mobile Web
<!-- Insert screenshots of your changes on the web platform (from a mobile browser)-->

https://user-images.githubusercontent.com/53711423/127908706-217507ec-28f3-405d-bd7c-c666816b2bad.mp4


#### Desktop
<!-- Insert screenshots of your changes on the desktop platform-->

https://user-images.githubusercontent.com/53711423/127908722-5d191741-0132-48d4-be57-c9b0091b16b3.mov



#### iOS
<!-- Insert screenshots of your changes on the iOS platform-->

https://user-images.githubusercontent.com/53711423/127908733-c0a3aacd-f77b-4a78-bb0b-59a6748a0a61.mp4



#### Android
<!-- Insert screenshots of your changes on the Android platform-->

https://user-images.githubusercontent.com/53711423/127909479-a6c34d3d-2571-49d3-a7cf-879935466148.mp4