@rdjuric 2021-08-02
No worries @marcaaron, I had forgot about Android 🥲 Should be ready for review now!