@marcaaron 2021-08-06
I don't think there are any specific steps besides regressions around navigating to a chat