@OSBotify 2021-08-11
🚀 [Deployed](https://github.com/Expensify/App/actions/runs/1120793764) to production by @francoisl in version: 1.0.83-1 🚀

 platform | result 
 ---|--- 
🤖 android 🤖|success ✅ 
🖥 desktop 🖥|failure ❌
🍎 iOS 🍎|success ✅ 
🕸 web 🕸|success ✅